Demo

+++ superset/custom_security_manager.py
@@ -1,4 +1,4 @@
-from flask import redirect, g, flash, request, session, current_app, url_for, make_response
+from flask import redirect, g, flash, request, session, current_app, url_for, make_response, Response
 from flask_appbuilder.security.views import UserDBModelView, AuthDBView,AuthLDAPView,AuthView,AuthOIDView
 from superset.security import SupersetSecurityManager
 from flask_appbuilder.security.views import expose
@@ -6,13 +6,14 @@ from flask_appbuilder.security.manager import BaseSecurityManager
 from flask_login import login_user, logout_user
 from flask_appbuilder.security.forms import LoginForm_db
 from urllib.parse import quote
+from flask_appbuilder._compat import as_unicode
 from superset import app
 import requests
+import json
+from urllib.parse import urlparse
+from posixpath import basename, dirname
 
-#config = app.config
-#sso_app_host = config["SSO_HOST"]
-#sso_app_port = config["SSO_PORT"] 
-#sso_app_name = config["SSO_NAME"]
+
 
 class SSOSessionClient:
     def __init__(self, sso_app_host, sso_app_port, sso_app_name):
@@ -28,39 +29,88 @@ class SSOSessionClient:
             files = {'applicationName': app_name, 'username': username}
             response = requests.post(self.get_sso_session_app_url(), files=files)
             response.raise_for_status()
-            return response.json()
+            return json.loads(response.text)
         except Exception as e:
             return str(e)
 
     def get_sso_session(self, sso_session_id):
+        sso = {'username': 'Nouser'}
         try:
-            response = requests.get(self.get_sso_session_app_url() + '/' + sso_session_id)
+            response = requests.get(self.get_sso_session_app_url() + '/' + sso_session_id )
             response.raise_for_status()
-            return response.json()
+            return json.loads(response.text)
         except Exception as e:
-            return str(e)
+            return sso
 
     def delete_sso_session(self, sso_session_id):
         try:
             response = requests.delete(self.get_sso_session_app_url() + '/' + sso_session_id)
             response.raise_for_status()
-            return response.json()
+            return json.loads(response.text)
         except Exception as e:
             return str(e)
+
+
 
 
 class CustomAuthDBView(AuthDBView):
     login_template = 'appbuilder/general/security/login_db.html'
 
-
+    
     @expose('/login/', methods=['GET', 'POST'])
     def login(self):
+
+        
+
+        """sso starts"""
+
+        sso_app_host = self.appbuilder.app.config["SSO_HOST"]
+        sso_app_port = self.appbuilder.app.config["SSO_PORT"] 
+        sso_app_name = self.appbuilder.app.config["SSO_NAME"]
+        client = SSOSessionClient(sso_app_host, sso_app_port , sso_app_name )
+
+        
+        name = request.cookies.get('sso')       
+ 
+        if name is not None:
+
+
+            client = SSOSessionClient(sso_app_host, sso_app_port , sso_app_name )
+            sso_id = client.get_sso_session(name)
+            username = sso_id['username']
+
+            user = self.appbuilder.sm.find_user(username=username)
+
+
+
+            if user is not None:
+                
+
+                                
+                login_user(user, remember=False)
+
+                arg = request.args
+                dashboard_id = request.args.get('next')
+
+               
+
+                if dashboard_id is not None: 
+                    parse_object = urlparse(dashboard_id)
+                    s=parse_object.path
+                    s=s.split('/')[-2] 
+                    return redirect("/superset/dashboard/{}".format(s))
+
+
+                return redirect("/superset/welcome")
+
+        """sso ends"""
+
 
         if g.user is not None and g.user.is_authenticated:
             return redirect(self.appbuilder.get_url_for_index)
         form = LoginForm_db()
         if form.validate_on_submit():
-            user = self.appbuilder.sm.auth_user_db(
+            user = self.appbuilder.sm.auth_user_ldap(
                 form.username.data, form.password.data
             )
             if not user:
@@ -68,17 +118,25 @@ class CustomAuthDBView(AuthDBView):
                 return redirect(self.appbuilder.get_url_for_login)
             login_user(user, remember=False)
 
+            sso_app_host = self.appbuilder.app.config["SSO_HOST"]
+            sso_app_port = self.appbuilder.app.config["SSO_PORT"] 
+            sso_app_name = self.appbuilder.app.config["SSO_NAME"]
+            sso_api_name = self.appbuilder.app.config["SSO_API_NAME"]
+            sso_domain = self.appbuilder.app.config["SSO_DOMAIN_NAME"]
+
             response = make_response(redirect(self.appbuilder.get_url_for_index))
-            client = SSOSessionClient('34.212.135.8', 1978, 'ssosession')
-            postResponse = client.create_sso_session('SUPERSET', user.username )
+            client = SSOSessionClient(sso_app_host, sso_app_port, sso_app_name )
+            postResponse = client.create_sso_session(sso_api_name , user.username )
             cook = postResponse['id']
-            #cook = user.username +':'+ form.password.data
+
 
 
             if response is not None:
-                #response.set_cookie('sso', value=cook )
-                response.set_cookie('sso', value=cook , domain = '.qubz-bi.com' )
+
+                response.set_cookie('sso', value=cook , domain = sso_domain )
                 return response 
+
+            
 
             return redirect(self.appbuilder.get_url_for_index)
 
@@ -88,10 +146,30 @@ class CustomAuthDBView(AuthDBView):
 
     @expose("/logout/")
     def logout(self):
+
+        sso_app_host = self.appbuilder.app.config["SSO_HOST"]
+        sso_app_port = self.appbuilder.app.config["SSO_PORT"]
+        sso_app_name = self.appbuilder.app.config["SSO_NAME"]
+        sso_domain = self.appbuilder.app.config["SSO_DOMAIN_NAME"]
+
+
         name = request.cookies.get('sso')
+        if name is not None:
+
+            client = SSOSessionClient(sso_app_host, sso_app_port, sso_app_name )
+            client.delete_sso_session(name)
+
+            res = app.make_response(redirect(self.appbuilder.get_url_for_index))
+    #
+
+            
+            res.set_cookie('sso', expires=0,path='/',domain=sso_domain)
+
+            logout_user()
+            return res
         logout_user()
-        client = SSOSessionClient('34.212.135.8', 1978, 'ssosession')
-        client.delete_sso_session(name)
+
+        
         return redirect(self.appbuilder.get_url_for_index)
 
 class CustomSecurityManager(SupersetSecurityManager):
@@ -99,3 +177,5 @@ class CustomSecurityManager(SupersetSecurityManager):
     #authdbview  = CustomAuthDBView
     def __init__(self, appbuilder):
         super(CustomSecurityManager, self).__init__(appbuilder)
+
+
